# eRegistry - Spring/ApiDocumentation.json
{
  "swagger": "2.0",
  "info": {
    "description": "Api Documentation",
    "version": "1.0",
    "title": "Api Documentation",
    "termsOfService": "urn:tos",
    "contact": {},
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "host": "localhost:8090",
  "basePath": "/",
...
  "paths": {
...
    "/People": {
      "get": {
        "tags": [
          "people-controller"
        ],
        "summary": "getAllPeople",
        "operationId": "getAllPeopleUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Serializable"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/People/Address/id={idAddress}": {
      "get": {
        "tags": [
          "people-controller"
        ],
        "summary": "getAddressById",
        "operationId": "getAddressByIdUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "idAddress",
            "in": "path",
            "description": "idAddress",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Serializable"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "people-controller"
        ],
        "summary": "removeAddressById",
        "operationId": "removeAddressByIdUsingDELETE",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "idAddress",
            "in": "path",
            "description": "idAddress",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Serializable"
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/People/Child/id={idPerson}": {
      "get": {
        "tags": [
          "people-controller"
        ],
        "summary": "getAllChild",
        "operationId": "getAllChildUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "idPerson",
            "in": "path",
            "description": "idPerson",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Serializable"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/People/Person/id={idPerson}": {
      "get": {
        "tags": [
          "people-controller"
        ],
        "summary": "getPersonById",
        "operationId": "getPersonByIdUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "idPerson",
            "in": "path",
            "description": "idPerson",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Serializable"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "people-controller"
        ],
        "summary": "removePersonById",
        "operationId": "removePersonByIdUsingDELETE",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "idPerson",
            "in": "path",
            "description": "idPerson",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Serializable"
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/People/newAddress": {
      "post": {
        "tags": [
          "people-controller"
        ],
        "summary": "insertAddress",
        "operationId": "insertAddressUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "address",
            "description": "address",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Address"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Serializable"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/People/newPerson": {
      "post": {
        "tags": [
          "people-controller"
        ],
        "summary": "insertPerson",
        "operationId": "insertPersonUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "person",
            "description": "person",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Person"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Serializable"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/People/newPersonWithAddress": {
      "post": {
        "tags": [
          "people-controller"
        ],
        "summary": "insertPersonWithAddress",
        "operationId": "insertPersonWithAddressUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "insertPersonWithAddress",
            "description": "insertPersonWithAddress",
            "required": true,
            "schema": {
              "$ref": "#/definitions/InsertPersonWithAddress"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Serializable"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/People/updateAddress": {
      "post": {
        "tags": [
          "people-controller"
        ],
        "summary": "updateAddress",
        "operationId": "updateAddressUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "updateAddressRequest",
            "description": "updateAddressRequest",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateAddressRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Serializable"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/People/updateExpirationDate": {
      "post": {
        "tags": [
          "people-controller"
        ],
        "summary": "updateExpirationDate",
        "operationId": "updateExpirationDateUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "updateExpirationDateRequest",
            "description": "updateExpirationDateRequest",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateExpirationDateRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Serializable"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/People/updateIdAddress": {
      "post": {
        "tags": [
          "people-controller"
        ],
        "summary": "updateIdAddress",
        "operationId": "updateIdAddressUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "updateIdAddressRequest",
            "description": "updateIdAddressRequest",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateIdAddressRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Serializable"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/People/updateMail": {
      "post": {
        "tags": [
          "people-controller"
        ],
        "summary": "updateMail",
        "operationId": "updateMailUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "updateMailRequest",
            "description": "updateMailRequest",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateMailRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Serializable"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/People/updatePhone": {
      "post": {
        "tags": [
          "people-controller"
        ],
        "summary": "updatePhone",
        "operationId": "updatePhoneUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "updatePhoneRequest",
            "description": "updatePhoneRequest",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdatePhoneRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Serializable"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Serializable": {
      "type": "object"
    },
    "Address": {
      "type": "object",
      "properties": {
        "city": {
          "type": "string"
        },
        "country": {
          "type": "string"
        },
        "flatNumber": {
          "type": "integer",
          "format": "int32"
        },
        "houseNumber": {
          "type": "string"
        },
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "postalCode": {
          "type": "string"
        },
        "street": {
          "type": "string"
        }
      }
    },
    "UpdateExpirationDateRequest": {
      "type": "object",
      "properties": {
        "expirationDate": {
          "type": "string"
        },
        "idPerson": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "InsertPersonWithAddress": {
      "type": "object",
      "properties": {
        "address": {
          "$ref": "#/definitions/Address"
        },
        "person": {
          "$ref": "#/definitions/Person"
        }
      }
    },
...
    "UpdateAddressRequest": {
      "type": "object",
      "properties": {
        "city": {
          "type": "string"
        },
        "country": {
          "type": "string"
        },
        "flatNumber": {
          "type": "integer",
          "format": "int32"
        },
        "houseNumber": {
          "type": "string"
        },
        "idAddress": {
          "type": "integer",
          "format": "int32"
        },
        "postalCode": {
          "type": "string"
        },
        "street": {
          "type": "string"
        }
      }
    },
    "UpdateMailRequest": {
      "type": "object",
      "properties": {
        "idPerson": {
          "type": "integer",
          "format": "int32"
        },
        "mail": {
          "type": "string"
        }
      }
    },
    "UpdatePhoneRequest": {
      "type": "object",
      "properties": {
        "idPerson": {
          "type": "integer",
          "format": "int32"
        },
        "phone": {
          "type": "string"
        }
      }
    },
...
    "UpdateIdAddressRequest": {
      "type": "object",
      "properties": {
        "idPerson": {
          "type": "integer",
          "format": "int32"
        },
        "newIdAddress": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "Person": {
      "type": "object",
      "properties": {
        "dateOfBirth": {
          "type": "string"
        },
        "expirationDate": {
          "type": "string"
        },
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "idAddress": {
          "type": "integer",
          "format": "int32"
        },
        "mail": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "sex": {
          "type": "string"
        },
        "surname": {
          "type": "string"
        }
      }
    },
...
  }
}